requirements.txt
flask==3.0.0
redis==5.0.1
chromadb==0.4.22
playwright==1.40.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
argparse
numpy==1.24.3
sentence-transformers==2.2.2
uvicorn==0.24.0
